Drop trailing slash from site URL
I noticed that the article suggested citations all have two slashes between the domain part and the path part of the URL.  I noticed the same thing in the <link> tag on line 5 of https://vsoch.github.io/feed.xml.

I think this is because those pages are forming the full URL using `{{ site.url }}{{ site.baseurl }}`, `site.baseurl` is absolute, and `site.url` ends in a slash.  So, this change resolves that by removing the trailing slash from the site URL!

The Jekyll docs page (https://jekyllrb.com/docs/variables/) also leaves the trailing slash out of the example.